+++ sickrage/providers/torrent/freshontv.py
@@ -66,11 +66,11 @@ class FreshOnTVProvider(TorrentProvider):
         return True
 
     def login(self):
-        if any(requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.sess.cookies).values()):
+        if any(requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.cookies).values()):
             return True
 
         if self._uid and self._hash:
-            requests.utils.add_dict_to_cookiejar(sickrage.srCore.srWebSession.sess.cookies, self.cookies)
+            requests.utils.add_dict_to_cookiejar(sickrage.srCore.srWebSession.cookies, self.cookies)
         else:
             login_params = {'username': self.username,
                             'password': self.password,
@@ -85,10 +85,10 @@ class FreshOnTVProvider(TorrentProvider):
             if re.search('/logout.php', response):
 
                 try:
-                    if requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.sess.cookies)['uid'] and \
-                            requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.sess.cookies)['pass']:
-                        self._uid = requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.sess.cookies)['uid']
-                        self._hash = requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.sess.cookies)['pass']
+                    if requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.cookies)['uid'] and \
+                            requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.cookies)['pass']:
+                        self._uid = requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.cookies)['uid']
+                        self._hash = requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.cookies)['pass']
 
                         self.cookies = {'uid': self._uid,
                                         'pass': self._hash}