+++ sickrage/providers/torrent/hdtorrents.py
@@ -65,7 +65,7 @@ class HDTorrentsProvider(TorrentProvider):
 
     def login(self):
 
-        if any(requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.sess.cookies).values()):
+        if any(requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.cookies).values()):
             return True
 
         login_params = {'uid': self.username,